обновление

+++ src/main/java/ru/programming/problems/problemseven/ArrayPI.java
@@ -1,5 +1,5 @@
 package ru.programming.problems.problemseven;
 
 public class ArrayPI {
-    public int[] array = new int[35];
+    public long[] array = new long[35];
 }
